+++ mitra-json-signatures/src/create.rs
@@ -26,6 +26,7 @@ use mitra_utils::{
 
 use super::proofs::{
     CRYPTOSUITE_JCS_EDDSA,
+    CRYPTOSUITE_JCS_EDDSA_LEGACY,
     DATA_INTEGRITY_PROOF,
     PROOF_TYPE_JCS_BLAKE2_ED25519,
     PROOF_TYPE_JCS_EIP191,
@@ -64,6 +65,19 @@ impl IntegrityProofConfig {
         Self {
             proof_type: DATA_INTEGRITY_PROOF.to_string(),
             cryptosuite: Some(CRYPTOSUITE_JCS_EDDSA.to_string()),
+            proof_purpose: PURPOSE_ASSERTION_METHOD.to_string(),
+            verification_method: verification_method.to_string(),
+            created: created_at,
+        }
+    }
+
+    pub fn jcs_eddsa_legacy(
+        verification_method: &str,
+        created_at: DateTime<Utc>,
+    ) -> Self {
+        Self {
+            proof_type: DATA_INTEGRITY_PROOF.to_string(),
+            cryptosuite: Some(CRYPTOSUITE_JCS_EDDSA_LEGACY.to_string()),
             proof_purpose: PURPOSE_ASSERTION_METHOD.to_string(),
             verification_method: verification_method.to_string(),
             created: created_at,