+++ src/mastodon_api/accounts/views.rs
@@ -437,7 +437,7 @@ async fn get_identity_claim(
         "minisign-unhashed" => {
             let did_key = minisign_key_to_did(&query_params.signer)
                 .map_err(|_| ValidationError("invalid key"))?;
-            (Did::Key(did_key), IdentityProofType::FepC390JcsEddsaProof)
+            (Did::Key(did_key), IdentityProofType::LegacyFepC390JcsEddsaProof)
         },
         _ => return Err(ValidationError("unknown proof type").into()),
     };
@@ -469,7 +469,7 @@ async fn create_identity_proof(
     let proof_type = match proof_data.proof_type.as_str() {
         "ethereum" => IdentityProofType::FepC390JcsEip191Proof,
         "minisign" => IdentityProofType::FepC390JcsBlake2Ed25519Proof,
-        "minisign-unhashed" => IdentityProofType::FepC390JcsEddsaProof,
+        "minisign-unhashed" => IdentityProofType::LegacyFepC390JcsEddsaProof,
         _ => return Err(ValidationError("unknown proof type").into()),
     };
     let did = proof_data.did.parse::<Did>()
@@ -540,7 +540,7 @@ async fn create_identity_proof(
             ).map_err(|_| ValidationError("invalid signature"))?;
             signature_bin
         },
-        IdentityProofType::FepC390JcsEddsaProof => {
+        IdentityProofType::LegacyFepC390JcsEddsaProof => {
             let did_key = did.as_did_key()
                 .ok_or(ValidationError("unexpected DID type"))?;
             let ed25519_key_bytes = did_key.try_ed25519_key()
@@ -552,7 +552,7 @@ async fn create_identity_proof(
             if signature.is_prehashed {
                 return Err(ValidationError("invalid signature type").into());
             };
-            let proof_config = IntegrityProofConfig::jcs_eddsa(
+            let proof_config = IntegrityProofConfig::jcs_eddsa_legacy(
                 &did_key.to_string(),
                 proof_data.created_at,
             );